refactor: establish frontend project structure with organized directories
- Create organized directory structure following React best practices
- Add components/, pages/, hooks/, utils/, services/, types/, assets/ directories
- Move App component to components/ directory with proper imports
- Add index files to each directory for clean exports
- Update import paths to maintain functionality
- Maintain existing assets and styling structure
