+++ frontend/src/components/common/index.ts
@@ -1,0 +1,2 @@
+// Common reusable components will be exported from here
+// Example: export { Button } from './Button';
