+++ frontend/src/components/features/index.ts
@@ -1,0 +1,2 @@
+// Feature-specific components will be exported from here
+// Example: export { UserProfile } from './UserProfile';
