+++ frontend/src/components/index.ts
@@ -1,0 +1,8 @@
+// Main App component
+export { default as App } from './App';
+
+// Common components
+export * from './common';
+
+// Feature-specific components
+export * from './features';
