+++ frontend/src/hooks/index.ts
@@ -1,0 +1,2 @@
+// Custom hooks will be exported from here
+// Example: export { useAuth } from './useAuth';
