+++ frontend/src/pages/index.ts
@@ -1,0 +1,2 @@
+// Page components will be exported from here
+// Example: export { HomePage } from './HomePage';
