+++ frontend/src/services/index.ts
@@ -1,0 +1,2 @@
+// API services and external integrations will be exported from here
+// Example: export { apiClient } from './apiClient';
