+++ frontend/src/types/index.ts
@@ -1,0 +1,2 @@
+// TypeScript type definitions will be exported from here
+// Example: export type { User, ApiResponse } from './api';
