+++ frontend/src/utils/index.ts
@@ -1,0 +1,2 @@
+// Utility functions will be exported from here
+// Example: export { formatDate } from './dateUtils';
